+++ internal/provider/provider.go
@@ -18,6 +18,7 @@ import (
 	"github.com/platform9/terraform-provider-pcd/internal/services/keymanager"
 	"github.com/platform9/terraform-provider-pcd/internal/services/loadbalancer"
 	"github.com/platform9/terraform-provider-pcd/internal/services/networking"
+	"github.com/platform9/terraform-provider-pcd/internal/services/vpnaas"
 )
 
 // Ensure pcdProvider satisfies the provider.Provider interface.
@@ -87,6 +88,11 @@ func (p *pcdProvider) Resources(_ context.Context) []func() resource.Resource {
 		dns.NewRecordSetResource,
 		keymanager.NewSecretResource,
 		keymanager.NewContainerResource,
+		vpnaas.NewServiceResource,
+		vpnaas.NewIKEPolicyResource,
+		vpnaas.NewIPSecPolicyResource,
+		vpnaas.NewEndpointGroupResource,
+		vpnaas.NewSiteConnectionResource,
 	}
 }